requirements.txt
uvicorn
fastapi
python-dotenv
colorama
ollama
chromadb



# installed but not used
# pydantic
# openai

